Configure ESLint for TypeScript
For now, we need to use the Git version of typescript-eslint-parser
because the last release only supports TypeScript 3.1. A new release
is planned for any day now, so we should be able to move to an NPM
release soon.

https://github.com/eslint/typescript-eslint-parser/pull/596

As ESLint's no-unused-vars doesn't work well with TypeScript, I
enabled noUnusedLocals and noUnusedParameters on the TypeScript
side instead.