TypeScript rewrite (#73)
* Update to Babel 7

This is a pre-requisite to using @babel/preset-typescript later.

This change required upgrading rollup-plugin-babel, which in turn
required upgrading Rollup.js itself. Fortunately, the upgraded
rollup-plugin-babel automatically takes care of Babel helpers
deduplication and turning off module transpilation in Babel's env
preset, so the Babel configuration could be drastically simplified.
For the same reason, we don't need rollup-plugin-replace anymore.

* Remove unused @babel/plugin-external-helpers

* Add Babel Typescript preset and port isPlainObject

* Port createAction() to TypeScript

* Port configureStore() to TypeScript

* Fix Rollup configuration

* Fix type errors in serializableStateInvariantMiddleware.test.ts

* Port createReducer() to TypeScript

* Port createSlice() to TypeScript

* Port index.js to TS and add index.d.ts

* Add some typings-tester tests

* Add "target" to tsconfig.json

It's apparently not strictly needed for Babel's TypeScript plugin,
but at least more explicit.

* Use `T extends string` instaead of `string` as action type

This allows more concrete types like `Action<'increment'>` to
be inferred by TypeScript.

* Remove unneeded extension from Rollup input path

* Remove unneeded type parameter

* Avoid a type cast in createReducer()

* Type `isPlainObject` argument as `unknown`

* Make findNonSerializableValue() typing more precise

* Remove redundant "lib" in tsconfig.json

It's implied by "target": "es2018".

* Enable "esModuleInterop" in tsconfig.json

* Update typings-tester tsconfig.json

* Fix configureStore() state type inference edge case

When passing an object as `reducer` and at the same time
`preloadedState`, TypeScript infers the store state type from the
latter although `preloadedState` is potentially just a partial state.
An additional `PS` type parameter fixes this.

* Fix configureStore() state inference without extra type

Thanks to @Dudeonyx for this hack.

* Introduce EnhancedStore type for thunk dispatch

* Update formatting commands

* Reformat, again

* Package lock random changes

* Conditionally import immutable-state-invariant middleware

* Make createReducer() return Reducer<Immutable<S>, A>

This reflects the actual runtime behavior of immer (which freezes
produced values) and fixes a type error with immer >= 1.10.3.

* Revert createReducer() type change

Things are fixed in v1.10.5.

* 0.4.0-0

* Run TypeScript compiler in `prepare`

* Configure ESLint for TypeScript

For now, we need to use the Git version of typescript-eslint-parser
because the last release only supports TypeScript 3.1. A new release
is planned for any day now, so we should be able to move to an NPM
release soon.

https://github.com/eslint/typescript-eslint-parser/pull/596

As ESLint's no-unused-vars doesn't work well with TypeScript, I
enabled noUnusedLocals and noUnusedParameters on the TypeScript
side instead.

* Generate TypeScript declaration files in dist/

* Revert "Generate TypeScript declaration files in dist/"

This reverts commit e449c8a78a501920c0d96705f81b46f8870e4e6a.

* Publish src/ for TypeScript typings

* 0.4.0-1