fix: use correct working directory for CLI test
GitHub actions override the working directory in tests, which causes
`path.resolve` to point to nonexisting files. There are some mentions of
similar issues on SO.

This fix resolves paths from `process.env.PWD` which points to the
process working directory which is set correctly both on local runs and
GitHub actions.