Fix a bug in the RTKQ selector typings causing bad state args
The `ReducerPath` type tries to extract the generic type for "the
path of the reducer" arg from `EndpointDefinitions`. However, it
looks like this was always off by one: getting the 4th arg instead
of the actual 5th arg.

This caused bad input to the generated selectors. That seems to
have worked okay with Reselect 4.0 and its looser types, but
caused problems when we updated to Reselect 4.1.x and its improved
types. The `RootState` type collapsed down to an empty object,
because there was no valid string to use as an object key.

This fixes the bug and adds a typetest to verify that the selector
types are carried through all the way.