Rewrite subscription handling to batch actions
When a component mounts and tries to subscribe, it dispatches a thunk.
The first thunk dispatched for a given cache entry makes the actual
request. Later thunks for the same cache entry will see that the cache
entry exists, return `false` from `condition()`, and bail out.  That
dispatches a `rejected` action, and the reducer logic listened for that
to add the additional subscription entry.

However, this meant that adding N subscriptions to a cache entry would
dispatch N + 1: `pending` + `fulfilled` from the first thunk, and N-1
`rejected` actions from all others.

We've had reports that this led to slow perf when rendering large lists
of components with query hooks inside.

I've rewritten the logic to catch all `query/rejected` subscription
actions within an event loop tick, and dispatch them as a single
batched action at the end of the tick with `queueMicrotask`. This
drastically minimizes the number of dispatched actions - for a list of
1000 items, it went from 1002 actions to 4, and 6.5s to 1s.