Upgrade version of "graphql-request"
Upgrading the version of "graphql-request" to use the latest. This is
not only to get onto the latest version of the package, but also to
remove an internal bug to that can surface from it. In this case, there
are duplicate `AbortSignal` types, one defined by the Typescript
language, and one defined by the package. This update removes this
conflict.